Initial interface int -> long conversion

+++ src/ServiceStack.Interfaces/Redis/IRedisHash.cs
@@ -8,6 +8,6 @@ namespace ServiceStack.Redis
 	{
 		bool AddIfNotExists(KeyValuePair<string, string> item);
 		void AddRange(IEnumerable<KeyValuePair<string, string>> items);
-		int IncrementValue(string key, int incrementBy);
+		long IncrementValue(string key, int incrementBy);
 	}
 }
